@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -29,8 +29,8 @@ dependencies:
   equatable: ^2.0.3
   flutter_bloc: ^7.0.1
   http: ^0.13.3
-  dartz: ^0.9.2
-  get_it: ^7.1.0
+  dartz: ^0.10.0-nullsafety.2
+  get_it: ^7.1.3
   page_transition: ^2.0.2
 
 dev_dependencies:
