@@ -1,17 +1,17 @@
 import 'package:flutter/widgets.dart';
 
 class SizeConfig {
-  static  MediaQueryData _mediaQueryData;
-  static  double screenWidth;
-  static  double screenHeight;
-  static  double blockSizeHorizontal;
-  static  double blockSizeVertical;
+  static  late MediaQueryData _mediaQueryData;
+  static  double? screenWidth;
+  static  double? screenHeight;
+  static  late double blockSizeHorizontal;
+  static  late double blockSizeVertical;
 
   void init(BuildContext context) {
     _mediaQueryData = MediaQuery.of(context);
     screenWidth = _mediaQueryData.size.width;
     screenHeight = _mediaQueryData.size.height;
-    blockSizeHorizontal = screenWidth / 100;
-    blockSizeVertical = screenHeight / 100;
+    blockSizeHorizontal = screenWidth! / 100;
+    blockSizeVertical = screenHeight! / 100;
   }
 }
