@@ -4,7 +4,7 @@ part of 'class_room_page_bloc.dart';
 abstract class ClassRoomPageState extends Equatable{
 
   @override
-  List<Object> get props =>[];
+  List<Object?> get props =>[];
 }
 
 class Empty extends ClassRoomPageState {}
@@ -12,21 +12,21 @@ class Empty extends ClassRoomPageState {}
 class Loading extends ClassRoomPageState {}
 
 class Loaded extends ClassRoomPageState {
-  final ClassRoomsDataModel classrooms;
-  final ClassRoomDataModel classroom;
+  final ClassRoomsDataModel? classrooms;
+  final ClassRoomDataModel? classroom;
 
   Loaded({this.classrooms,this.classroom, });
 
   @override
-  List<Object> get props => [classrooms];
+  List<Object?> get props => [classrooms];
 
 }
 
 class Error extends ClassRoomPageState {
-  final String message;
+  final String? message;
 
   Error({this.message});
 
   @override
-  List<Object> get props => [message];
+  List<Object?> get props => [message];
 }
