@@ -4,7 +4,7 @@ part of 'inner_class_room_page_bloc.dart';
 abstract class InnerClassRoomPageState extends Equatable{
 
   @override
-  List<Object> get props =>[];
+  List<Object?> get props =>[];
 }
 
 class Empty extends InnerClassRoomPageState {}
@@ -13,20 +13,20 @@ class Loading extends InnerClassRoomPageState {}
 
 class Loaded extends InnerClassRoomPageState {
 
-  final ClassRoomDataModel classroom;
+  final ClassRoomDataModel? classroom;
 
   Loaded({this.classroom, });
 
   @override
-  List<Object> get props => [classroom];
+  List<Object?> get props => [classroom];
 
 }
 
 class Error extends InnerClassRoomPageState {
-  final String message;
+  final String? message;
 
   Error({this.message});
 
   @override
-  List<Object> get props => [message];
+  List<Object?> get props => [message];
 }
