@@ -11,20 +11,20 @@ class Loading extends RegistrationPageState {}
 class Empty extends RegistrationPageState {}
 
 class Error extends RegistrationPageState {
-  final String message;
+  final String? message;
   Error({this.message});
 }
 
 class Loaded extends RegistrationPageState {
 
-  final int code;
-  final RegistrationsDataModel registrationsDataModel;
+  final int? code;
+  final RegistrationsDataModel? registrationsDataModel;
   Loaded({ this.code, this.registrationsDataModel});
 }
 
 class Deleted extends RegistrationPageState{
   final int code;
 
-  Deleted({@required this.code});
+  Deleted({required this.code});
 
 }
