@@ -16,10 +16,10 @@ part 'student_subject_registration_state.dart';
 const String SERVER_FAILURE_MESSAGE = 'Server Failure';
 
 class StudentSubjectRegistrationBloc extends Bloc<StudentSubjectRegistrationEvent, StudentSubjectRegistrationState> {
-  final RegistrationsSetRegistrationUseCase setRegistrationUseCase;
-  final del.RegistrationsDeleteRegistrationUseCase deleteRegistrationUseCase;
-  final RegistrationsGetRegistrationsUseCase getRegistrationsUseCase;
-  StudentSubjectRegistrationBloc({@required RegistrationsSetRegistrationUseCase setRegistration,@required del.RegistrationsDeleteRegistrationUseCase deleteRegistration,@required  RegistrationsGetRegistrationsUseCase getRegistrations}) :setRegistrationUseCase=setRegistration,getRegistrationsUseCase=getRegistrations,deleteRegistrationUseCase=deleteRegistration, super(Empty());
+  final RegistrationsSetRegistrationUseCase? setRegistrationUseCase;
+  final del.RegistrationsDeleteRegistrationUseCase? deleteRegistrationUseCase;
+  final RegistrationsGetRegistrationsUseCase? getRegistrationsUseCase;
+  StudentSubjectRegistrationBloc({required RegistrationsSetRegistrationUseCase? setRegistration,required del.RegistrationsDeleteRegistrationUseCase? deleteRegistration,required  RegistrationsGetRegistrationsUseCase? getRegistrations}) :setRegistrationUseCase=setRegistration,getRegistrationsUseCase=getRegistrations,deleteRegistrationUseCase=deleteRegistration, super(Empty());
 
   StudentSubjectRegistrationState get initial=>Empty();
 
@@ -27,10 +27,10 @@ class StudentSubjectRegistrationBloc extends Bloc<StudentSubjectRegistrationEven
   Stream<StudentSubjectRegistrationState> mapEventToState(StudentSubjectRegistrationEvent event) async* {
 
     if(event is RegisterStudentSubject){
-      Failure failure;
+      Failure? failure;
       yield Loading();
       try{
-        final failureOrClassRoom= await setRegistrationUseCase(Params(subjectId: event.subjectId, studentId: event.studentId));
+        final failureOrClassRoom= await setRegistrationUseCase!(Params(subjectId: event.subjectId!, studentId: event.studentId!));
         failureOrClassRoom.fold((l) => failure=l, (r) => null);
         print("map event to state called");
         yield* _eitherLoadedOrErrorState2(failureOrClassRoom);
@@ -38,10 +38,10 @@ class StudentSubjectRegistrationBloc extends Bloc<StudentSubjectRegistrationEven
         yield Error(message: _mapFailureToMessage(failure));
       }
     }else if(event is DeleteRegistration){
-      Failure failure;
+      Failure? failure;
       yield Loading();
       try{
-        final failureOrClassRoom= await deleteRegistrationUseCase(del.Params(registrationId: event.registrationId));
+        final failureOrClassRoom= await deleteRegistrationUseCase!(del.Params(registrationId: event.registrationId!));
         failureOrClassRoom.fold((l) => failure=l, (r) => null);
         print("map event to state called");
         yield* _eitherLoadedOrErrorState2(failureOrClassRoom);
@@ -50,10 +50,10 @@ class StudentSubjectRegistrationBloc extends Bloc<StudentSubjectRegistrationEven
       }
 
     }else if(event is GetRegistrations){
-      Failure failure;
+      Failure? failure;
       yield Loading();
       try{
-        final failureOrClassRoom= await getRegistrationsUseCase(NoParams());
+        final failureOrClassRoom= await getRegistrationsUseCase!(NoParams());
         failureOrClassRoom.fold((l) => failure=l, (r) => null);
         print("map event to state called");
         yield* _eitherLoadedOrErrorState(failureOrClassRoom);
@@ -80,7 +80,7 @@ class StudentSubjectRegistrationBloc extends Bloc<StudentSubjectRegistrationEven
   }
 
 
-  String _mapFailureToMessage(Failure failure) {
+  String _mapFailureToMessage(Failure? failure) {
 
     print("map failure called");
     switch (failure.runtimeType) {
