@@ -12,20 +12,20 @@ class Loading extends StudentSubjectRegistrationState {}
 class Empty extends StudentSubjectRegistrationState {}
 
 class Error extends StudentSubjectRegistrationState {
-  final String message;
+  final String? message;
   Error({this.message});
 }
 
 class Loaded extends StudentSubjectRegistrationState {
 
-  final int code;
-  final RegistrationsDataModel registrationsDataModel;
+  final int? code;
+  final RegistrationsDataModel? registrationsDataModel;
   Loaded({ this.code, this.registrationsDataModel});
 }
 
 class Deleted extends StudentSubjectRegistrationState{
   final int code;
 
-  Deleted({@required this.code});
+  Deleted({required this.code});
 
 }
